// src/lib/api/generated/models/problemWithDetailsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Campus Ativo API
 * API para gerenciamento de problemas de infraestrutura
 * OpenAPI spec version: 1.0.0
 */
import type { AttachmentInfo } from './attachmentInfo';
import type { ProblemReporterResponse } from './problemReporterResponse';
import type { ProblemWithDetailsResponseDeletedAt } from './problemWithDetailsResponseDeletedAt';
import type { ProblemWithDetailsResponseStatus } from './problemWithDetailsResponseStatus';

export interface ProblemWithDetailsResponse {
  /** ID do problema */
  id: string;
  /** Dados do autor do problema */
  reporter: ProblemReporterResponse;
  /** Título do problema */
  title: string;
  /** Slug do problema */
  slug: string;
  /** Trecho da descrição do problema (primeiros 120 caracteres) */
  excerpt: string;
  /** Nome da localização do problema */
  locationName: string;
  /** Status do problema */
  status: ProblemWithDetailsResponseStatus;
  /** Data de criação */
  createdAt: string;
  /** Data de atualização */
  updatedAt?: string;
  /**
   * Data de exclusão (null se não deletado)
   * @nullable
   */
  deletedAt?: ProblemWithDetailsResponseDeletedAt;
  /** Anexos do problema */
  attachments?: AttachmentInfo[];
}